CIV-14827 and 28, test fix
